requirements.txt
alabaster==0.7.12
attrs==21.2.0
awscli==1.19.59
Babel==2.9.0
backcall==0.2.0
botocore==1.20.59
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.5
Cython==0.29.23
decorator==5.0.7
docutils==0.15.2
flake8==3.9.1
gensim==4.0.1
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
loguru==0.5.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
smart-open==5.1.0
snowballstemmer==2.1.0
sparsesvd==0.2.2
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
-e git+git@github.com:av-maslov/solid-disco.git@6ed85c3d400bc7746adfbd8e5eeeed8d5b742e34#egg=src
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tqdm==4.61.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
